Add loop logic
