+++ src/AISmart.Application.Contracts/Agents/MockA/Events/AThreadCreatedEvent.cs
