+++ src/AISmart.Application.Contracts/Agents/MockB/Events/BThreadCreatedEvent.cs
