+++ src/AISmart.Application.Contracts/Agents/MockC/Events/CThreadCreatedEvent.cs
