+++ src/AISmart.Domain.Shared/Agents/MockA/MockAAgentState.cs
@@ -1,10 +1,10 @@
 using System.Collections.Generic;
 using Orleans;
 
-namespace AISmart.Agents.A;
+namespace AISmart.Agents.MockA;
 
 [GenerateSerializer]
-public class AAgentState : StateBase
+public class MockAAgentState : StateBase
 {
     [Id(0)] public List<string> ThreadIds { get; set; }
     [Id(1)] public int Number { get; set; }
