+++ src/AISmart.Domain.Shared/Agents/MockA/MockAGEvent.cs
