+++ src/AISmart.Domain.Shared/Agents/MockB/MockBAgentState.cs
