+++ src/AISmart.Domain.Shared/Agents/MockB/MockBGEvent.cs
