+++ src/AISmart.Domain.Shared/Agents/MockC/CAgentState.cs
