+++ src/AISmart.Domain.Shared/Agents/MockC/CGEvent.cs
