+++ test/AISmart.GAgents.Tests/LoadTest/LoadTest.cs
@@ -1,58 +1,48 @@
-using AISmart.Agents.A.Events;
-using AISmart.Application.Grains.Agents.A;
-using AISmart.Application.Grains.Agents.B;
 using AISmart.Application.Grains.Agents.Group;
 using AISmart.Application.Grains.Agents.Publisher;
-using AISmart.Application.Grains.Agents.C;
 using AISmart.GGrains.Tests;
-using Xunit.Abstractions;
-using System.Diagnostics;
+using AISmart.Agents.MockA.Events;
+using AISmart.Application.Grains.Agents.MockA;
+using AISmart.Application.Grains.Agents.MockB;
+using AISmart.Application.Grains.Agents.MockC;
 
 namespace AISmart.Grains.Tests.LoadTest;
 
-public class PerformanceTest : GAgentTestKitBase
+public class LoadTest : GAgentTestKitBase
 {
-    private readonly ITestOutputHelper _testOutputHelper;
-
-    public PerformanceTest(ITestOutputHelper testOutputHelper)
+    [Theory]
+    [InlineData(100, 100, 100)]
+    public async Task AgentLoadTest(int aGAgentCount, int bGAgentCount, int cGAgentCount)
     {
-        _testOutputHelper = testOutputHelper;
-    }
-
-    [Fact]
-    public async Task LoadTest()
-    {
-        var stopwatch = new Stopwatch();
-        stopwatch.Start();
-        _testOutputHelper.WriteLine($"Test started at: {DateTime.Now}");
-
         var groupGAgent = await Silo.CreateGrainAsync<GroupGAgent>(Guid.NewGuid());
-        var aGAgent = await Silo.CreateGrainAsync<AGAgent>(Guid.NewGuid());
-        var bGAgent = await Silo.CreateGrainAsync<BGAgent>(Guid.NewGuid());
-        var cGAgent = await Silo.CreateGrainAsync<CGAgent>(Guid.NewGuid());
-
         var publishingAgent = await Silo.CreateGrainAsync<PublishingGAgent>(Guid.NewGuid());
-        await groupGAgent.Register(aGAgent);
-        await groupGAgent.Register(bGAgent);
-        await groupGAgent.Register(cGAgent);
+
+        for (int i = 0; i < aGAgentCount; i++)
+        {
+            var aGAgent = await Silo.CreateGrainAsync<MockAGAgent>(Guid.NewGuid());
+
+            await groupGAgent.Register(aGAgent);
+        }
+
+        for (int i = 0; i < bGAgentCount; i++)
+        {
+            var bGAgent = await Silo.CreateGrainAsync<MockBGAgent>(Guid.NewGuid());
+
+            await groupGAgent.Register(bGAgent);
+        }
+
+        for (int i = 0; i < cGAgentCount; i++)
+        {
+            var cGAgent = await Silo.CreateGrainAsync<MockCGAgent>(Guid.NewGuid());
+            await groupGAgent.Register(cGAgent);
+        }
+
         await publishingAgent.PublishTo(groupGAgent);
 
-        await publishingAgent.PublishEventAsync(new AThreadCreatedEvent
+        await publishingAgent.PublishEventAsync(new MockAThreadCreatedEvent
         {
             Id = $"mock_A_thread_id",
             Content = $"Testing iteration"
         });
-        
-        var aAgentState = await aGAgent.GetStateAsync();
-        var bAgentState = await bGAgent.GetStateAsync();
-        var cAgentState = await cGAgent.GetStateAsync();
-        
-        _testOutputHelper.WriteLine($"AGAgent State: {aAgentState.Number}");
-        _testOutputHelper.WriteLine($"BGAgent State: {bAgentState.Number}");
-        _testOutputHelper.WriteLine($"CGAgent State: {cAgentState.Number}");
-        
-        stopwatch.Stop();
-        _testOutputHelper.WriteLine($"Test ended at: {DateTime.Now}");
-        _testOutputHelper.WriteLine($"Total execution time: {stopwatch.ElapsedMilliseconds} ms");
     }
 }
